Ahom-Full Stack/backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-environ==0.11.2
psycopg2-binary==2.9.10
gunicorn==23.0.0
requests==2.32.3
stripe==11.4.1
django-ratelimit==4.1.0
Pillow==11.1.0
